# perf: remove repeated fee accumulator unmarshaling in swaps