<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰    
v    Before smashing the submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
v    If your PR doesn't close an issue, that's OK!  Just remove the Closes: #XXX line!
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

Closes: #XXX

## What is the purpose of the change

This is a third improvement identified in:
https://github.com/osmosis-labs/osmosis/pull/5287

Thi improvement focuses on removing the need for refetching and reunmarshaking fee accumulator on every swap step.

We only need to fetch it once, get its value every time a tick is crossed and update at the very end of the swap.

Drive-by changes:
- remove awkward abstraction called `ChargeFee`. I think it is clearer to add to the accumulator directly
- replace `ChargeFee` with `AddToFeeAccumulator` helper in tests

Since this is the last improvement identified, this is the flame graph as of [this commit]([116170d](https://github.com/osmosis-labs/osmosis/pull/5298/commits/116170d047f825c6ba7dedf6a827eeb587d39ae8))

![image](https://github.com/osmosis-labs/osmosis/assets/34196718/7599f059-dd92-4375-b573-1e1605e42866)

Comparison to commit from yesterday: https://github.com/osmosis-labs/osmosis/commit/8b523efcb4db53e4a74756405e6e5253d657089c

```
root@roman:~/osmosis/x/concentrated-liquidity# benchstat current.bench latest.bench
current.bench:5: missing iteration count
latest.bench:5: missing iteration count
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/osmosis/v15/x/concentrated-liquidity
cpu: Intel(R) Xeon(R) Gold 6248 CPU @ 2.50GHz
                    │ current.bench │            latest.bench            │
                    │    sec/op     │   sec/op     vs base               │
SwapExactAmountIn-4    341.1m ± 14%   173.9m ± 8%  -49.02% (p=0.002 n=6)
```


## Testing and Verifying

This change is already covered by existing tests

## Documentation and Release Note

  - [ ] Does this pull request introduce a new feature or user-facing behavior changes?
  - [ ] Changelog entry added to `Unreleased` section of `CHANGELOG.md`?

Where is the change documented? 
  - [ ] Specification (`x/{module}/README.md`)
  - [ ] Osmosis documentation site
  - [ ] Code comments?
  - [x] N/A